+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 authclient extension Change Log
 2.2.18 under development
 ------------------------
 
+- Bug #396: Fix `BaseOAuth` method - `refreshAccessToken` when no refresh token exists (kalmer)
 - Bug #393: Fix type for `BaseOAuth` property - `accessToken` (max-s-lab)